Rename Rialto module to Rialto-PoA
This will reduce confusion as the Millau runtime has a
Rialto module as well which refers to the Substrate chain.
